pyproject.toml
[project]
name = "zen-claw-ai"
version = "0.1.3.post5"
description = "A light and safe personal AI assistant framework"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "zen-claw contributors"}
]
keywords = ["ai", "agent", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "litellm>=1.81.10,<2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "httpx[socks]>=0.25.0",
    "loguru>=0.7.0",
    "readability-lxml>=0.8.0",
    "rich>=13.0.0",
    "croniter>=2.0.0",
    "python-telegram-bot[socks]>=21.0",
    "lark-oapi>=1.0.0",
    "socksio>=1.0.0",
    "cryptography>=42.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
rag = [
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "trafilatura>=1.9.0",
    "jieba>=0.42.1",
    "rank-bm25>=0.2.2",
]
tts = [
    "edge-tts>=6.1.9",
]
tts-all = [
    "edge-tts>=6.1.9",
    "httpx>=0.27.0",
]
multitenant = [
    "PyJWT>=2.8.0",
    "bcrypt>=4.0.0",
]

[project.scripts]
zen-claw = "zen_claw.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["zen_claw"]

[tool.hatch.build.targets.wheel.sources]
"zen_claw" = "zen_claw"

# Include non-Python files in skills
[tool.hatch.build]
include = [
    "zen_claw/**/*.py",
    "zen_claw/skills/**/*.md",
    "zen_claw/skills/**/*.sh",
]

[tool.hatch.build.targets.sdist]
include = [
    "zen_claw/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "zen_claw/bridge"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
resolution = "highest"
python-preference = "managed"
dev-dependencies = []

